chore(api-headless-cms-pg-os): split abstractions into separate files
One abstraction per file: EntryWriteOperations.ts and EntrySearchOperations.ts
in their own abstractions/ directory.
